switch SensorData to short[ ] since input values aren't expected beyond +-32768 range, and it double storage capacity for longer intervals

+++ LinearEncoder.h
@@ -24,7 +24,7 @@ class LinearEncoder : public SensorInput {
 			SensorInput::label = "LinearEncoder";
 		}
 
-		int poll() {
+		short poll() {
 			#ifdef DEBUG
 				if (Serial) {
 					Serial.println("LinearEncoder::poll()");
